Add reload! as a IRB::HelperMethod so it can be combined with other calls in a single statement. Fixes rails/rails#52982
It is useful to iterate and test in the console by reloading and executing something in a single line command. Something like:

reload!; Post.find(123).test_a_thing

This makes it easy to quickly switch to the console, hit up and enter to re-execute after a code change.

This restores functionality that used to work until a few months ago.