Update Guide on Multi-Database wrt DATABASE_URL
The Guide on Multiple Databases has a NOTE added that addresses a potential gotcha (thatgotma’) where migrating to a multi-database setup means that the DATABASE_URL configuration method that is often used in simpler application stops working entirely.

The solution to this is shown if the user wants to keep using that method with an explicit `url:` configuration example.

Hopefully this will help out the next person from not being surprised!