Include nested strict loaded associations on reload
If strict loading is enabled rails preloads associations when calling
`reload`. However, this did not include nested associations that can be
preloaded by using a Hash. For example when using

```ruby
Post.includes(comments: :user)
```

the User record would not have been preloaded when reloading a Post.
With this change we recurse the association cache to include those
nested associations.

[Fix #53017]