Explicitly set the user to vscode in Dev Container
The ruby image installs ruby for the vscode user (which is created in the debian base image). Vscode uses this user by default. However, other dev container implementations may not do that, in which case the dev container may open as root where ruby is not available. So let's explicitly set the user in the devcontainer.json.