Remove "Suppressing Callbacks" section from ActiveRecord guide
The recent documentation overhauls is great, but unfortunately the section on suppressing callbacks is wrong. Despite what is implied by the short description in the API docs, ActiveRecord::Base.suppress does not actually interact with callbacks per-se, but prevents records of the suppressed class from being saved at all. Because of this, the example given in the guide is very incorrect.