ConnectionPool#unpin_connection! handle lazily pinned connections
Fix: https://github.com/rails/rails/issues/53147

In `setup_transactional_fixtures` we register a callback to lazily
pin a connections in pools that are created after the fixtures setup.

But when this callback is used, the connection owner might be a different
thread than the one that will call `teardown_transactional_fixtures`.

To handle that, we must steal the connection when doing the final
unpinning.