Fix log level of :retry_stopped event
`subscribe_log_level` was not called for `:retry_stopped` event but for
`:enqueue_retry` event. So no log level was set for `:retry_stopped`
event, and the log level of `:enqueue_retry` event was altered from
`:info` to `:error`.

This commit fixes the log level of `:retry_stopped` event to `:error`
and restores the log level of `:enqueue_retry` event to `:info`.