Set Current.session if signed in in unauthenticated access
Currently, if you call `allow_unauthenticated_access` in a controller,
`Current.session` won't be set. This can result in rendered views making
the user appear to be signed out if they are already signed in.

This updates `allow_unauthenticated_access` to set a `before_action` of
`resume_session` to allow the user to still be signed in for requests
that do not require authentication.