Fix infinite loop in ERBTracker
When we nest strings inside ERBTracker our earlier regex only finds a
partial match on the string and then it's possible to have a mismatched
unmber of brackets. scan_until does not advance the string when it does
not find a match.

Ultimately, we can't parse Ruby via regex, so this is a best effort. We
do need to make sure to fail safely.

Co-authored-by: Breno Gazzola <breno.gazzola@gmail.com>