Fix timezone for records created using `.insert` and `.insert_all` in MySQL database
According to MySQL's documentation, `CURRENT_TIMESTAMP()` is a synonym
for `NOW()` and:

> Returns the current date and time as a value in `'YYYY-MM-DD hh:mm:ss'
> (...) The value is expressed in the session time zone.

Also:

> MySQL converts `TIMESTAMP` values from the current time zone to UTC for
> storage, and back from UTC to the current time zone for retrieval. (This
> does not occur for other types such as `DATETIME`.) By default, the
> current time zone for each connection is the server's time. The time
> zone can be set on a per-connection basis. As long as the time zone
> setting remains constant, you get back the same value you store. If you
> store a `TIMESTAMP` value, and then change the time zone and retrieve the
> value, the retrieved value is different from the value you stored.

This results in the `created_at` and `updated_at` timestamps for records
created with `.insert` and `.insert_all` to be offset by as many hours
as the server's timezone.

Most servers in production run in UTC, so that is not a problem, but
when developing on a local machine, the timezone is usually the one for
where one lives.

I live in Japan and records created with `.insert` get a `created_at`
that's 9 hours in the future.

I'm not sure how to write tests in the test suite for this, but you'll
find more details and a way to reproduce in this issue: #53250.