Refactor `activerecord/Rakefile` and add `port` option to the config
My first need was to change the port used to connect to my local MySQL
server when running commands like `bundle exec rake db:mysql:build`.

As I worked my way through that, I realized I could refactor make the
code in the Rakefile DRYer, which is the second part of this commit.