Fix unintended guide chapter scrolling when visit by turbo [skip ci]
When visiting a guide page by turbo, the navHighlight function fires
the scrollIntoView API.
Then, the chapter scrolls to the end of the chapter list by the
updateHighlight function because IntersectionObserver's time origin
is the first page loaded.

To fix this, add "is the page navigated by directly" to the condition
of the call updateHighlight function.