Flaky test demo
Seems like we have a flaky test in AR.  This commit just demonstrates
the problem.  I'm not sure what to do about it.  AFAICT, we're caching
the default time zone in the TypeMap on ActiveRecord.  Presumably
`with_timezone_config` should clear that cache, but I'm not sure the
right way to do it.

This commit just calls `test_default_in_utc` to demonstrate the problem
and make the test consistently fail