Prevent `ActiveSupport::TimeZone` documentation class inferences
Prior to this change, the API documentation transformed the mention of
"Time" into a link to the [Time][] class, and incorrectly inferred the
mention of captial M "Method" as a mention of Rails' [Method][] class
extensions.

After this change, the tokens are treated as capitalized English words
instead of Ruby Class names.

[Time]: https://api.rubyonrails.org/classes/Time.html
[Method]: https://api.rubyonrails.org/classes/Method.html