Fix authentication_generator_test
Those tests started to fail as soon as we started developing 8.1 since
it is not present on rubygems.

We should not have to generate a new app and install gems to be able to
test the behavior of this generator.

I used the same techniques we use in the application generator mocking
the commands that are executed and generating only the files that are
necessary to test the behavior of this generator.