Translate `no connection to the server` to ConnectionNotEstablished
Fixes https://github.com/rails/rails/issues/53399

In PostgreSQL, if a network error occurs when submitting a query to the server,
the `PG::UnableToSend: no connection to the server` error may occur.

- https://github.com/ged/ruby-pg/blob/v1.5.8/ext/pg_connection.c#L1945
- https://github.com/postgres/postgres/blob/REL_17_0/src/interfaces/libpq/fe-exec.c#L1689

The current low-level exception translator `translate_exception` primarily translates
`PG::ConnectionBad` into a retryable exception, so exceptions like the one above
will not be retried by `allow_retry: true`.

However, as we can see from the libpq, this query is guaranteed to return an error
before being executed, so it should be safe to retry it.