Exercise option name labels test
Originally this did test that if single keyword arguments is recognized
as label values not options since label values are required arguments.

It works fine when two label values and keyword arguments are passed,
but if users are unsure whether we can use a label with the same name as
an option at the same time, I'll add an extra test.