Enable schema queries to be retryable
When `permanent_connection_checkout` was [introduced][1], Active Record
was changed to always use `with_connection` and [never][2]
`lease_connection`. This causes apps that never call `lease_connection`
on their own (or in a gem) during a request to no longer have a
connection pinned to the current Thread/Fiber. By checking in/out
connections as needed, the number of times a connection must be verified
during a single request has significantly increased (more pings = more
latency).

The good news is that Active Record also now [defers][3] connection
verification and can use retryable queries as verification instead of
explicit pings.

This commit marks most internal SCHEMA queries as retryable so that they
can be used as connection verification. While this likely won't help as
much in production for applications using a schema cache dump, it will
definitely decrease latency for applications in development and those
that do not use a dump in production.

(There are a few places where `query_value{,s}` modifies database state,
those were not marked as retryable by this commit)

[1]: 4db9f512591b0188f05c0762102a3386b43c1b34
[2]: 7c68c5210cbc245d778daa7958cab73bc74f4669
[3]: 7fe221d898afe877f282b549bd8bb908ebe1843d