Reset Deduplicable singleton caches in schema cache tests
The deduplication code can mask bugs -- if a new field gets added to the
object but without it being added to the hash/equal methods then cache
loading will appear to just work in tests, but in reality be broken.

Here we reset the dedup registries before testing cache loading.