Fix routes with `::` in the path
When a route has `::` in the path, the scanner was not able to parse it
correctly. This was because the scanner was not checking if the next
byte was a `:` when the current byte was a `:`.

This will match the behavior of the old parser.

It feels like the `#peek_byte` on the string scanner should
accept a position to peek at, but it doesn't. So we have to use
`#getbyte` to get the byte at the next position.

Fixes #53397.