Unify query values deduplication for multi values
In the multi values normalization #39333, I also wanted to deduplicate
that (#38145, 29874cc), but the lack of deduplication for some multi
values was leaking into the behavior (#39171), so we had to deprecate
that behavior first (#39358).

And then the newly added multi values caused an issue again due to the
lack of deduplication (#53468).

To avoid such issues in the future, deduplicate all remaining multi
values that are not currently deduplicated, and add a test to ensure
that all multi values (including those that will be added in the future)
are deduplicated.