Introduce a basic skeleton for "database maintenance" operations
Most RDBMSes have some way to update the query planner by analyzing
the data and indexes. Rails currently does not provide an easy way for
developers or operators to do this.

For example:

- in PostgreSQL, the `ANALYZE` statement:
  - https://www.postgresql.org/docs/current/sql-analyze.html
- in MySQL/MariaDB, the `ANALYZE TABLE` statement:
  - https://dev.mysql.com/doc/refman/8.4/en/analyze-table.html
  - https://mariadb.com/kb/en/analyze-table/
- in SQLite3, the `PRAGMA optimize` and `ANALYZE` statements:
  - https://www.sqlite.org/pragma.html#pragma_optimize
  - https://www.sqlite.org/lang_analyze.html

This commit introduces a new rake tasks:

- db:maintenance:analyze

This task is intended to update query planner statistics in a way that
will not interfere with a running application. Typically this might
mean using the `SKIP_LOCKED` parameter in PostgreSQL's `ANALYZE`
statement, or setting the `analysis_limit` bit in SQLite3's `PRAGMA
optimize` statement.

The actual implementation for these operations should be in the
relevant connection adapters, by implementing a method

> `def perform_maintenance(event)`

where `event` is a symbol indicating the type of maintenance that
should be done.

Right now, the only option invoked by rails is `:analyze`, but it's
intended to be extensible to support more operations (e.g., "vacuum"
or "reindex"). I can even imagine connection adapters potentially
subscribing to events that should trigger different kinds of
maintenance (such as `CREATE INDEX` or a large number of rows being
added or deleted).