Call out pitfalls with Encryption when upgrading
As has been noted in a few issues, upgrading from Rails to 7.1 config defaults
can lead to data loss or confusion as the hash_digest_class changed from SHA1 to
SHA256 but we don't by default class changed from SHA1 to SHA256 but we don't by
default class changed from SHA1 to SHA256 but we don't by default class changed
from SHA1 to SHA256 but we don't by default set
`support_sha1_for_non_deterministic_encryption = true`.

This adds the change to the 7.0->7.1 upgrade notes and to the 7.1 release notes.

https://github.com/rails/rails/issues/48204
https://github.com/rails/rails/issues/50212
https://github.com/rails/rails/issues/50226
https://calebhearth.com/rails-7-1-encryption-errors