Fix permissions error when building devcontainer
The apt-get step of the Dockerfile for the devcontainer started to fail
with a permissions error. This fixes that error by setting the user to
root during the execution of apt-get and subsequent COPY operations.
This should be low impact since it only happens in the context of the
devcontainer, and the user is set to vscode later in the process.