[Fix #47849] Correctly strip newlines from check constraints for MySQL 8.0.16+
The regular expression used to strip whitespace characters from check constraints that was added in #47851
was incorrectly specified, and resulted in `n` characters being left in the output.

For example, the constraint:

    json_schema_valid(_utf8mb4\'\\n        {\\n          "oneOf": [\\n            {\\n              "type": "null"\\n            },\\n            {\\n              "type": "array",\\n              "minItems": 1,\\n              "items": {\\n                "type": "integer",\\n                "minimum": 0\\n              }\\n            }\\n          ]\\n        }\',`booking_approver_ids`)

would result in the output:

    json_schema_valid(_utf8mb4'n {n "oneOf": [n {n "type": "null"n },n {n "type": "array",n "minItems": 1,n "items": {n "type": "integer",n "minimum": 0n }n }n ]n }',`booking_approver_ids`)

This is invalid when we restore the schema dump.

With this fix in place, the output becomes:

    json_schema_valid(_utf8mb4' { "oneOf": [ { "type": "null" }, { "type": "array", "minItems": 1, "items": { "type": "integer", "minimum": 0 } } ] }',`booking_approver_ids`)

This is valid and allows the schema dump to be restored correctly.

As a bonus, the revised code no longer uses regular expressions to avoid the same mistake occurring in the future.