Make column_definitions queries retryable
This commit follows up a [previous commit][1] that made internal schema
queries retryable. Previously, only the `PostgresqlAdapter` used `query`
(which is retryable by default) when fetching `column_definitions` to
populate the schema cache. This means applications without a schema
cache in production using MySQL would execute a `ping` query for each
table when loaded for the first time.

This commit adds `allow_retry: true` to `column_definitions` to prevent
needing to verify connections when executing these queries. `query` can
not be used (without a larger refactor) because it returns rows from the
`ActiveRecord::Result` instead of the `Result` itself.

[1]: b4c19d038e558d12c3e13317759ca13587c41a16