Set `framework_defaults` in all bug report templates
When testing bugs using the bug report templates, it helps if the
correct framework defaults are enabled for all templates.
This makes sure the scripts behave more inline with a regular Rails
application and allows easily testing older defaults.

This also normalizes the templates somewhat and removes some unneeded
code.