Support uploading a stream which is at end-of-stream
When Active Storage uploads a stream, it does the following:

  1. Compute a checksum of the stream[^1]
  2. Call the storage service to perform the actual upload
  3. Compare the checksum computed in step 1 with the checksum of
     the uploaded file and raise an `IntegrityError` if they don't
     match[^2]

The checksum in step 1 is computed from the current offset of the stream
but the stream is rewound after. Thus, the whole stream is uploaded in
step 2, regardless of the stream's initial offset. As a result, if the
stream has an initial offset, the integrity check in step 3 fails.

A few people have run into this issue[^3] and the root cause is not
immediately obvious from the error raised by Active Storage.

If the stream is at end-of-stream (i.e. if `io.eof?` is true), then the
user most likely expects the whole stream to be uploaded. So let's
rewind the stream in that case.

If the stream has an offset but is not at end-of-stream, then the user
could arguably expect the stream to be uploaded from the offset. In that
case, rewinding the stream would be surprising behaviour. Therefore,
let's preserve the existing behaviour in this scenario, i.e. raise an
`IntegrityError`. Ideally, we would bail out earlier with a more helpful
error but I wasn't sure where to raise so I'm leaving this as a
follow-up.

[^1]: https://github.com/rails/rails/blob/v8.0.0/activestorage/app/models/active_storage/blob.rb#L348-L359
[^2]: https://github.com/rails/rails/blob/v8.0.0/activestorage/lib/active_storage/service/disk_service.rb#L163-L168
[^3]: https://stackoverflow.com/questions/52335835/when-is-activestorageintegrityerror-raised