Fix with_connection not pinning when permanent
`ActiveRecord.permanent_connection_checkout` was [introduced][1] along
with `with_connection` to enable granular connection checkouts for
application that do not want connections pinned to a Thread/Fiber for
the duration of a request. The configuration was intended to be opt-in,
however when applications exclusively use `with_connection`
(not `{lease_,}connection`), connections will not be pinned regardless
of `permanent_connection_checkout`'s value.

This commit fixes the issue by ensuring connection checked out by
`with_connection` will be pinned when `permanent_connection_checkout` is
`true`.

Additionally, the private/nodoc `permanent_lease?` was renamed to
`sticky_unset?` because that more accurately reflects the method's
behavior and allows us to make assertions that the connection was
"permanently leased" by checking the sticky bit for truthiness.

[1]: dd8fd52c072b8f6777ad6d8acf86054fa1923dd9

Co-authored-by: zzak <zzakscott@gmail.com>