Active Storage: support disposition filename
While browsers support reading a filename for a downloaded file from an
`<a>` element's `[download]` attribute, that attribute's value has a
lower precedence than a filename specified in the response's
[Content-Disposition][] header.

When files are served with an `inline;`-prefixed `Content-Disposition`
header, the absence of a filename allows the browser to read the
filename from the `[download]` attribute on the `<a>` element.

However, files served with `attachment;`-prefixed `Content-Disposition`
headers (either explicitly through requests to to `rails_blob_url(…,
disposition: "attachment")` URLs or implicitly for files with types
listed in [config.active_storage.content_types_to_serve_as_binary][])
cannot control the name of the downloaded file. For most services, this
means that applications cannot override storage-service generated files
(for example:
`-var-folders-4q-q308vb3s12x2rzrgppdsz0540000gn-T-20241114-56163-dfyqb1.pdf`).

This commit adds support for overriding that filename with a `download`
query parameter (through the same mechanism as the [`disposition` query
parameter][redirect-mode]). The structure of the keyword arguments
nested underneath the `:disposition` key was chosen to match the
expected arguments for [ActionDispatch::Http::ContentDisposition][].

[download]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#download
[Content-Disposition]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition
[config.active_storage.content_types_to_serve_as_binary]: https://guides.rubyonrails.org/configuring.html#config-active-storage-content-types-to-serve-as-binary
[redirect-mode]: https://guides.rubyonrails.org/v8.0.0/active_storage_overview.html#redirect-mode
[ActionDispatch::Http::ContentDisposition]: https://github.com/rails/rails/blob/v8.0.3/actionpack/lib/action_dispatch/http/content_disposition.rb#L8