Maintain connection verification briefly after use
This means that a database connection that fails in between two requests
that are less than two seconds apart may cause the failure-following
request to die (if its first query is not retryable).

That's not a big concern in practice: per-request verification is
intended to protect against the case that the database connection failed
some large time before the request arrives. A request beginning within
seconds of the failure is morally equivalent to a request that was
already in flight.

Between automatically retryable select queries and BEGINs, the first
query of a request will very commonly be retryable anyway. This change's
value is in our new use of short-term leasing via with_connection, where
we can otherwise end up re-verifying for every non-retryable query.