Provide first-class support for patches
The changes in this commit allow developers to write patches to classes
in other gems without breaking lazy loading. With the previous method of
ignoring an `app/overrides` folder in the autoloader, and then `load`ing
all of the files in that directory in a `config.to_prepare` hook, all
the modified classes are implicitly eager loaded, which significantly
slows down app startup on complex apps.

Instead, this makes use of the `on_load` callback that `zeitwerk`
defines, and loads any modules that prepend themselves into loaded
classes after they are loaded.