Only link back to referer if referer is a `GET` request
Currently, passing `:back` to `link_to` will link to the `referer` if it
exists, regardless of its request method. This means there are cases
where we will link to a `referer` for a `POST` request, which results in
unexpected behavior. Take the following as an example.

```ruby
def create
  @product = Product.new(product_params)

  if @product.save
    redirect_to products_path, notice: "Product was successfully created."
  else
    render :new, status: :unprocessable_entity
  end
end
```

```erb
<% # app/views/products/new.html.erb %>

<%= link_to "Go back", :back %>
```

If a user were to navigate to `Products#new` from a page, and enter
invalid data, the server would respond with `Products#new`, making the
_that_ the `referer`.

This commit checks to see if the `"REQUEST_METHOD"` was `"GET"` before
setting the `referer`.