Add canonical link tag to guides head
The canonical link tells search engines what URL is preferred for the
current document. By point to the latest released version, it makes sure
search engines prefer the latest versions in search results.

Guides that have been removed will have a canonical URL that will result
in a 404. These will mostyl be ignored by search engines, but we can
also not show the canonical url for guides that will be deleted in the
next version.