Action Text: Dispatch Active Storage events with `id` and `file`
Follow-up to [#52680][]

Adjusts the Action Text-dispatched [Active Storage Direct Upload
JavaScript Events][] to include both `id` and `file` properties as
documented in the current [Action Text Attachment Direct Upload
JavaScript Events][]. In addition to those properties, there were two
events missing from the list:

* `direct-upload:before-blob-request`, `{id, file, xhr}`, Before making
  a request to your application for direct upload metadata.
* `direct-upload:before-storage-request`, `{id, file, xhr}`, Before
  making a request to store a file.

The code change mimics the style of the Active Storage
[DirectUploadController][] implementation.

Prior to this commit, the documentation mentioned that an `<input>`
element would be the event target. In reality, the `AttachmentUpload`
JavaScript class dispatches events treating the `<trix-editor>` element
as its target.

This commit changes the documentation to account for that. Alongside the
code change, this commit expands upon the existing guides documentation
to also mention the `attachment` property that references a
[Attachment][] instance.

To account for both _new events_ and _new properties_ in the
`event.detail` object, this commit includes expanded system test
coverage asserting that events are dispatched as described .

[#52680]: https://github.com/rails/rails/pull/52680
[Active Storage Direct Upload JavaScript Events]: https://guides.rubyonrails.org/v8.0.0/active_storage_overview.html#direct-upload-javascript-events
[Action Text Attachment Direct Upload JavaScript Events]: https://guides.rubyonrails.org/v8.0.0/action_text_overview.html#attachment-direct-upload-javascript-events
[DirectUploadController]: https://github.com/rails/rails/blob/v8.0.0/activestorage/app/javascript/activestorage/direct_upload_controller.js#L44-L48
[Attachment]: https://github.com/basecamp/trix/?tab=readme-ov-file#inserting-a-file