Support loading SQLite3 extensions with `config/database.yml`
The `sqlite3` gem v2.4.0 introduces support for loading extensions
that are passed to the Database constructor. This feature leverages
that feature to allow configuration of extensions using either
filesystem paths or the names of modules that respond to `.to_path`.

This commit documents the feature in both SQLite3Adapter rdoc and the
"Configuring" guide. It also extends and improves the documentation
around general SQLite3Adapter configuration.

See sparklemotion/sqlite3-ruby#586 for more information.